+++ src/next/jdbc.clj
@@ -332,7 +332,8 @@
            params)))
   ([connectable sql param-groups opts]
    (if (or (instance? java.sql.Connection connectable)
-           (opts/wrapped-connection? connectable))
+           (and (satisfies? p/Connectable connectable)
+                (instance? java.sql.Connection (:connectable connectable))))
      (with-open [ps (prepare connectable [sql] opts)]
        (execute-batch! ps param-groups opts))
      (with-open [con (get-connection connectable)]